vendoring: Add TUF and securesystemslib
Add the projects to vendoring config (but don't add the
vendored sources yet to keep the branch readable).

Expose USER_DATA_DIR to PipSession
This will be used to store the TUF metadata in a later commit.

network: Add secure_repository

LinkCollector: Use SecureRepository to download

Rename the fetch_page() to fetch_project_page() as it's
meant for specifically downloading project index pages.

Use SecureRepository to download the project page if we have one for
the remote repository. Otherwise download as before.

PipSession now holds the SecureRepositoryManager (PipSession does not
need it itself but the SecureRepositoryManager is used in the same
places as the PipSession...)

download: Use secure_repository to download distributions

network: Refactor download helpers

Make the progress notification decision into a function for
re-usability. Move some download helper functions into utils so they
can be used by code that does downloads for SecureRepository

secure_repository: Implement custom fetcher

This allows us to plug in the progress notifications (and to use the
same PipSession for SecureRepository downloads and normal ones).

secure_repository: Add secure metadata bootstrapping

Pip can ship known good metadata for repositories (pypi). Add the
infrastructure needed for this, and a bootstrap example.