LinkCollector: Use SecureRepository to download
Rename the fetch_page() to fetch_project_page() as it's
meant for specifically downloading project index pages.

Use SecureRepository to download the project page if we have one for
the remote repository. Otherwise download as before.

PipSession now holds the SecureRepositoryManager (PipSession does not
need it itself but the SecureRepositoryManager is used in the same
places as the PipSession...)